VirtualMemorySimulator v.0.7
Mostly everything is done. There is no code refactoring, it lacks of
software engineering and it does not have very useful comments. Apart
from that, the average access time and the hit ratio are the same in
the 1,000,000 trace. The bugs lays in the "Page Ins" & "Page Outs".
Please let me know if you discover what is going on.